Inline and simplify the After logic
Instead of using shared `handler` and `closer` variables and having most
of the common logic in `After`, move it into where it's used and remove
the After function and shared variables.

This makes the code only 1 line longer but makes it much easier to
follow. (and more similar to how the upstream code is, so makes merging
other changes easier)